+++ perf/inproc_thr.c
@@ -56,7 +56,7 @@ void worker (void *arg)
 
     for (i = 0; i != message_count; i++) {
         rc = nn_send (s, buf, message_size, 0);
-        assert (rc == message_size);
+        assert (rc == (int)message_size);
     }
 
     free (buf);
@@ -102,7 +102,7 @@ int main (int argc, char *argv [])
 
     for (i = 0; i != message_count; i++) {
         rc = nn_recv (s, buf, message_size, 0);
-        assert (rc == message_size);
+        assert (rc == (int)message_size);
     }
 
     elapsed = nn_stopwatch_term (&stopwatch);
